Refactor collection components (#14220)
* extract 'NormalItem' into own component

* extract CollectionCopyEntityModal

* extract CollectionSectionHeading

* extract CollectionEditMenu

* fix imports on CollectionEditMenu

* move pinned items to own component

* pass actions to pinned items

* pass component properly

* extract BulkActions

* move modals into bulk actions

* extract main collection items

* pass onMove and onCopy

* remove unused dashboard modal

* extract header into own component

* clean up imports

* organize imports

* fix sneaky dnd import
