+++ frontend/src/metabase/collections/constants.js
@@ -1,2 +1,3 @@
 // unit to use for calculating collection sidebar spacing (in px)
 export const SIDEBAR_SPACER = 18;
+export const ANALYTICS_CONTEXT = "Collection Landing";
